+++ clients/all.d.ts
@@ -376,3 +376,4 @@ export import SupplyChain = require('./supplychain');
 export import Artifact = require('./artifact');
 export import Chatbot = require('./chatbot');
 export import TimestreamInfluxDB = require('./timestreaminfluxdb');
+export import CodeConnections = require('./codeconnections');